GrpcFullBackEnd/service2/requirements.txt
psycopg2-binary==2.9.3
redis==4.2.2
grpcio
grpcio-tools